Added challenge feature (see description)
Challenge feature: In a standard UNO game, you can only play a draw 4 wild card if it is the only card you can play. If you choose to play a draw 4 wild card, the next player can then choose to challenge your decision to see if the draw 4 card was your only available card. If the draw 4 card was indeed the only available card, then you are safe and the challenger must draw 6 instead of drawing 4. However, if you are caught, you must draw 4 instead. It isn't necessary to challenge, though, and you can still decide to just regularly draw 4.